feat: updates CodeBlock and AboutPage
