+++ src/app/about/page.tsx
@@ -1,3 +1,4 @@
+import CodeBlock from "@/components/CodeBlock";
 import AboutBento from "@/components/pages/about/AboutBento";
 
 const AboutPage = () => {
@@ -8,6 +9,8 @@ const AboutPage = () => {
       </h1>
 
       <section className="mt-10">
+        <CodeBlock />
+        <div className="mt-24"></div>
         <AboutBento />
       </section>
     </div>
